% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOI_data.R
\docType{data}
\name{GOI_data}
\alias{GOI_data}
\title{Literacy Rates in India (GOI Data)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 35 rows and 8 columns.
}
\source{
\url{https://www.kaggle.com/datasets/doncorleone92/govt-of-india-literacy-rate?resource=download}
}
\usage{
GOI_data
}
\description{
A data set containing literacy rates across different states and union territories in India,
with data being from the year 2001 and 2011 census.
}
\details{
#' @format A data frame with 36 rows and 8 variables.
\describe{
\item{Category}{Category of the region (e.g., Country, State, Union Territory)}
\item{Country/States/Union Territories Name}{Name of the country, state, or union territory}
\item{Literacy Rate (Persons) - Total - 2001}{Overall literacy rate for 2001 (percentage)}
\item{Literacy Rate (Persons) - Total - 2011}{Overall literacy rate for 2011 (percentage)}
\item{Literacy Rate (Persons) - Rural - 2001}{Rural literacy rate for 2001 (percentage)}
\item{Literacy Rate (Persons) - Rural - 2011}{Rural literacy rate for 2011 (percentage)}
\item{Literacy Rate (Persons) - Urban - 2001}{Urban literacy rate for 2001 (percentage)}
\item{Literacy Rate (Persons) - Urban - 2011}{Urban literacy rate for 2011 (percentage)}
}
}
\keyword{datasets}
