% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literacy_growth.R
\name{literacy_growth}
\alias{literacy_growth}
\title{Calculate Literacy Growth (2001-2011)}
\usage{
literacy_growth(data)
}
\arguments{
\item{data}{A data frame containing literacy rates from the GOI dataset.}
}
\value{
A data frame with the original data and an additional column, \code{literacy_growth}, showing the growth rate.
}
\description{
This function calculates the percentage growth in literacy rates
from 2001 to 2011 for each state or union territory in the dataset.
}
\examples{
literacy_growth(data = GOI_data)
}
